chore: break integration tests up, make AENS test use random names
